+++ Service/ToolbarServiceInterface.php
@@ -1,0 +1,24 @@
+<?php
+/*
+ * (c) 2018: 975L <contact@975l.com>
+ * (c) 2018: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\ToolbarBundle\Service;
+
+/**
+ * Interface to be called for DI for Payment Main related services
+ * @author Laurent Marquet <laurent.marquet@laposte.net>
+ * @copyright 2018 975L <contact@975l.com>
+ */
+interface ToolbarServiceInterface
+{
+    /**
+     * Defines the icon and style for a define button
+     * @return array
+     */
+    public function defineButton(string $button);
+}